+++ packages/react-bindings/src/hooks/callUtilHooks.ts
@@ -7,6 +7,11 @@ import {
   hasScreenShareAudio,
   hasVideo,
   StreamVideoParticipant,
+  Call,
+  combineComparators,
+  screenSharing,
+  speakerLayoutSortPreset,
+  Comparator,
 } from '@stream-io/video-client';
 
 /**
@@ -156,4 +161,29 @@ export const useIsScreenShareAudioConnecting = (
     participant.screenShareAudioStream?.getAudioTracks()[0],
     hasScreenShareAudio(participant),
   );
+};
+
+const loggedIn: Comparator<StreamVideoParticipant> = (a, b) => {
+  if (a.isLocalParticipant) return 1;
+  if (b.isLocalParticipant) return -1;
+  return 0;
+};
+
+export const useSpeakerLayoutSortPreset = (
+  call: Call | undefined,
+  isOneOnOneCall: boolean,
+) => {
+  useEffect(() => {
+    if (!call) return;
+    const previousSortPreset = call.getSortParticipantsBy();
+    // always show the remote participant in the spotlight
+    if (isOneOnOneCall) {
+      call.setSortParticipantsBy(combineComparators(screenSharing, loggedIn));
+    } else {
+      call.setSortParticipantsBy(speakerLayoutSortPreset);
+    }
+    return () => {
+      call.setSortParticipantsBy(previousSortPreset);
+    };
+  }, [call, isOneOnOneCall]);
 };
